Update bibliographyController.js
Uses the helper function to remove arrays from the query and also change stepSize to -1 if it exists in query. Before any other number crashed the app because it was not being parsed as an int. However, this should be set to -1 first as the code doesn't handle other stepSizes at the moment.